fix: pt2 rename DataLoaderCache to fix case-sensitivity
