+++ packages/server/dataloader/RootDataLoader.ts
@@ -3,7 +3,7 @@ import {Logger} from '../utils/Logger'
 import * as atlassianLoaders from './atlassianLoaders'
 import * as azureDevOpsLoaders from './azureDevOpsLoaders'
 import * as customLoaderMakers from './customLoaderMakers'
-import DataLoaderCache from './DataLoaderCache2'
+import DataLoaderCache from './DataLoaderCache'
 import * as foreignKeyLoaderMakers from './foreignKeyLoaderMakers'
 import * as gcalLoaders from './gcalLoaders'
 import * as githubLoaders from './githubLoaders'
